plugin: patch node-pre-gyp module path variables
Patches node-pre-gyp's binary module_path substitutable variables
with fixed values in the native modules' package.json, so that cross
compiled native modules will be built in the same path where they
are expected to be found at runtime by node-pre-gyp. This is done
with a hook after cordova prepare, so it is applied after the
node project is copied for the target platform.
Also removes .gz files from the Android assets, for improved module
compatibility.