android: use helper script to call npm on macOS
On macOS, if Android Studio is not started through the command line,
the PATH environment variable will probably not contain npm's
location due to macOS environment variables in Windowed Applications
not being the same. This commit adds a cordova hook to create a
helper script that calls npm after adding the PATH elements that are
available during the Cordova project's prepare step, so that npm can
be called when the Gradle build is started by Android Studio.