bridge: emit message event inside a setImmediate
Emit the bridge's message event inside a setImmediate, so listener
code is called outside the napi scope created when the bridge
receives a call from the cordova application code.
This allows for promises started on the listeners to be resolved and
causes errors to be recognized by the javascript engine instead of
silently failing.