ios: cache native modules override preference
Saves the NODEJS_MOBILE_BUILD_NATIVE_MODULES environment variable
value to a file if it is defined during the Cordova prepare phase to
be read by iOS during build time, similar to the way it is working
for Android. It's also possible to create the file in www/ instead
of using the environment variable.